requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# AI/ML Libraries
litellm==1.17.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Graph and Network Analysis
networkx==3.2.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
pyvis==0.3.2
wordcloud==1.9.3

# Data Processing
numpy==1.26.2
pandas==2.1.4

# Utilities
Werkzeug==3.0.1
